Moved get_client to scope
