+++ sentry_sdk/client.py
@@ -167,6 +167,10 @@ class NoopClient:
         # type: () -> bool
         return False
 
+    def is_active(self):
+        # type: () -> bool
+        return False
+
     def __init__(self, *args, **kwargs):
         # type: (*Any, **Any) -> None
         return None
@@ -264,28 +268,6 @@ class _Client(NoopClient):
         # type: (Any) -> None
         self.options = state["options"]
         self._init_impl()
-
-    @classmethod
-    def get_client(cls):
-        # type: () -> Union[Client, NoopClient]
-        """
-        Returns the current :py:class:`sentry_sdk.Client`. If no client is available a :py:class:`sentry_sdk.client.NoopClient` is returned.
-
-        .. versionadded:: 1.XX.0
-        """
-        client = Scope.get_current_scope().client
-        if client is not None:
-            return client
-
-        client = Scope.get_isolation_scope().client
-        if client is not None:
-            return client
-
-        client = Scope.get_global_scope().client
-        if client is not None:
-            return client
-
-        return NoopClient()
 
     def _setup_instrumentation(self, functions_to_trace):
         # type: (Sequence[Dict[str, str]]) -> None
@@ -408,6 +390,10 @@ class _Client(NoopClient):
             _client_init_debug.set(old_debug)
 
         self._setup_instrumentation(self.options.get("functions_to_trace", []))
+
+    def is_active(self):
+        # type: () -> bool
+        return True
 
     @property
     def dsn(self):
